Add files via upload
This notebook implements a Deep Convolutional Generative Adversarial Network (DCGAN) to generate synthetic images of handwritten digits from the MNIST dataset. It features a generator that creates images from random noise and a discriminator that distinguishes between real and fake images. The model demonstrates how DCGANs, using deep convolutional networks, can be effectively trained to generate high-quality images. The project is built using TensorFlow and Keras, and the results showcase the generator's ability to mimic the distribution of handwritten digits in the MNIST dataset.